pyproject.toml
[tool.poetry]
name = "mot"
version = "0.2.0"
description = "A template for the cloud deployment of a labscript instance."
authors = ["fretchen <fred.jendrzejewski@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.6"
numpy = "^1.24.3"
lyse = "^3.2.3"
runmanager = "^3.2.0"
sqooler = {git = "https://github.com/Alqor-UG/sqooler.git"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pylint = "^3.0.3"
black = "^23.7.0"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
